+++ R/api_file_info.R
@@ -205,7 +205,7 @@ NULL
 
 .fi_during <- function(fi, start_date, end_date) {
     fi_tl <- .fi_timeline(fi)
-    .fi_switch(
+   .fi_switch(
         fi = fi,
         eo_cube = .between(fi_tl, start_date[[1]], end_date[[1]]),
         derived_cube = all(.between(fi_tl, start_date[[1]], end_date[[1]]))