+++ tests/testthat/test-cube.R
@@ -5,7 +5,6 @@ test_that("List collections", {
     expect_true(grepl("LANDSAT", col))
     expect_true(grepl("BDC", col))
 })
-
 test_that("api_source", {
     res_s2_b8a <- .source_bands_resolution(
         source = "MPC",
@@ -43,22 +42,20 @@ test_that("api_source", {
 
     open_bdc <- .source_collection_open_data(
         source = "BDC",
-        collection = "S2-SEN2COR_10_16D_STK-1"
+        collection = "SENTINEL-2-16D"
     )
     expect_true(open_bdc)
 
     token_bdc <- .source_collection_open_data(
         source = "BDC",
-        collection = "S2-SEN2COR_10_16D_STK-1",
+        collection = "SENTINEL-2-16D",
         token = TRUE
     )
     expect_true(token_bdc)
 
 })
-
 test_that("Reading a raster cube", {
     data_dir <- system.file("extdata/raster/mod13q1", package = "sits")
-
     raster_cube <- sits_cube(
             source = "BDC",
             collection = "MOD13Q1-6",
@@ -66,18 +63,16 @@ test_that("Reading a raster cube", {
             multicores = 2,
             progress = FALSE
     )
-
     # get bands names
     bands <- sits_bands(raster_cube)
     expect_true(all(bands %in% c("NDVI", "EVI")))
-
+    # test params
     params <- .raster_params_file(raster_cube$file_info[[1]]$path)
     expect_true(params$nrows == 147)
     expect_true(params$ncols == 255)
     expect_true(params$xres >= 231.5)
-
+    # test timeline
     timeline <- sits_timeline(raster_cube)
-
     sub_cube <- sits_select(raster_cube,
                             start_date = timeline[1],
                             end_date = timeline[2])
@@ -87,74 +82,92 @@ test_that("Reading a raster cube", {
     expect_true(params_2$ncols == 255)
     expect_true(params_2$xres >= 231.5)
 })
-
-test_that("Creating cubes from BDC", {
-
-
+test_that("Creating cubes from BDC - CBERS-WFI-16D", {
     # check "BDC_ACCESS_KEY" - mandatory one per user
     bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
-
     testthat::skip_if(nchar(bdc_access_key) == 0,
         message = "No BDC_ACCESS_KEY defined in environment."
     )
-
-    # create a raster cube file based on the information about the file
-    cbers_cube <- sits_cube(
-        source = "BDC",
-        collection = "CBERS-WFI-16D",
-        tiles = c("007004", "007005"),
-        start_date = "2018-09-01",
-        end_date = "2019-08-29",
-        progress = FALSE
-    )
-
-    testthat::skip_if(purrr::is_null(cbers_cube),
+    tiles <- c("007004", "007005")
+    start_date <- "2021-09-01"
+    end_date <- "2021-09-30"
+    bands <- c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")
+    # create a raster cube
+    cbers_cube_16d <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "CBERS-WFI-16D",
+            tiles = tiles,
+            start_date = start_date,
+            end_date = end_date,
+            progress = FALSE)
+    },
+    .default = NULL)
+    testthat::skip_if(purrr::is_null(cbers_cube_16d),
         message = "BDC is not accessible"
     )
-
-    expect_true(all(sits_bands(cbers_cube) %in%
-        c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")))
-    bbox <- sits_bbox(cbers_cube)
-    int_bbox <- .bbox_intersection(bbox, .tile_bbox(cbers_cube))
-    expect_true(all(int_bbox == sits_bbox(.tile(cbers_cube))))
-
-    timeline <- sits_timeline(cbers_cube)
-    expect_true(timeline[1] <= as.Date("2018-09-01"))
-    expect_true(timeline[length(timeline)] <= as.Date("2019-08-29"))
-
-    r_obj <- .raster_open_rast(cbers_cube$file_info[[1]]$path[1])
-    cube_nrows <- .tile_nrows(cbers_cube)
+    # test bands and bbox
+    expect_true(all(sits_bands(cbers_cube_16d) %in% bands))
+    bbox <- sits_bbox(cbers_cube_16d)
+    int_bbox <- .bbox_intersection(bbox, .tile_bbox(cbers_cube_16d))
+    expect_true(all(int_bbox == sits_bbox(.tile(cbers_cube_16d))))
+    # test timeline
+    timeline <- sits_timeline(cbers_cube_16d)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+    # test raster obj
+    r_obj <- .raster_open_rast(cbers_cube_16d$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(cbers_cube_16d)
     expect_true(.raster_nrows(r_obj) == cube_nrows)
-
-    fi_1 <- cbers_cube$file_info[[1]]
-    cc <- .fi_cloud_cover(fi_1)
-    expect_equal(length(cc), nrow(fi_1))
-    expect_equal(cc[[1]], 0.00)
-
-    fi_2 <- .fi_filter_fid(fi_1,
-            fid = "CB4-16D_V2_007004_20180829")
-    expect_equal(length(sits_bands(cbers_cube)), nrow(fi_2))
-
-    cc_2 <- .fi_cloud_cover(fi_2)
-    expect_true(all(cc_2 %in% c(0.00)))
 })
-
-test_that("Creating cubes from BDC - based on ROI with shapefile", {
-
-
+test_that("Creating cubes from BDC - CBERS-WFI-8D", {
     # check "BDC_ACCESS_KEY" - mandatory one per user
     bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
+    testthat::skip_if(nchar(bdc_access_key) == 0,
+                      message = "No BDC_ACCESS_KEY defined in environment."
+    )
+    tiles <- c("007004", "007005")
+    start_date <- "2022-05-01"
+    end_date <- "2022-08-29"
+    bands <- c("NDVI", "EVI", "B13", "B14", "B15", "B16", "CLOUD")
+    # create a raster cube file from BDC
+    cbers_cube_8d <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "CBERS-WFI-8D",
+            tiles = tiles,
+            start_date = start_date,
+            end_date = end_date,
+            progress = FALSE)
+    },
+    .default = NULL)
+    testthat::skip_if(purrr::is_null(cbers_cube_8d),
+                      message = "BDC is not accessible"
+    )
+    expect_true(all(sits_bands(cbers_cube_8d) %in% bands))
+    bbox <- sits_bbox(cbers_cube_8d)
+    int_bbox <- .bbox_intersection(bbox, .tile_bbox(cbers_cube_8d))
+    expect_true(all(int_bbox == sits_bbox(.tile(cbers_cube_8d))))
 
+    timeline <- sits_timeline(cbers_cube_8d)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+
+    r_obj <- .raster_open_rast(cbers_cube_8d$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(cbers_cube_8d)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating cubes from BDC - based on ROI with shapefile", {
+    # check "BDC_ACCESS_KEY" - mandatory one per user
+    bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
     testthat::skip_if(nchar(bdc_access_key) == 0,
         message = "No BDC_ACCESS_KEY defined in environment."
     )
-
     shp_file <- system.file(
         "extdata/shapefiles/mato_grosso/mt.shp",
         package = "sits"
     )
     sf_mt <- sf::read_sf(shp_file)
-
     # create a raster cube file based on the information about the files
     modis_cube <- .try({
         sits_cube(
@@ -168,33 +181,27 @@ test_that("Creating cubes from BDC - based on ROI with shapefile", {
         )
     },
     .default = NULL)
-
     testthat::skip_if(purrr::is_null(modis_cube),
         message = "BDC is not accessible"
     )
-
     expect_true(all(sits_bands(modis_cube) %in% c("NDVI", "EVI")))
     bbox <- sits_bbox(modis_cube, as_crs = "EPSG:4326")
     bbox_shp <- sf::st_bbox(sf_mt)
-
     expect_lt(bbox["xmin"], bbox_shp["xmin"])
     expect_lt(bbox["ymin"], bbox_shp["ymin"])
     expect_gt(bbox["xmax"], bbox_shp["xmax"])
     expect_gt(bbox["ymax"], bbox_shp["ymax"])
     intersects <- .cube_intersects(modis_cube, sf_mt)
     expect_true(all(intersects))
+
 })
-
 test_that("Creating cubes from BDC - invalid roi", {
-
-
     # check "BDC_ACCESS_KEY" - mandatory one per user
     bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
 
     testthat::skip_if(nchar(bdc_access_key) == 0,
         message = "No BDC_ACCESS_KEY defined in environment."
     )
-
     expect_error(
         object = sits_cube(
             source = "BDC",
@@ -206,7 +213,6 @@ test_that("Creating cubes from BDC - invalid roi", {
             progress = FALSE
         )
     )
-
     expect_error(
         object = sits_cube(
             source = "BDC",
@@ -225,9 +231,164 @@ test_that("Creating cubes from BDC - invalid roi", {
         )
     )
 })
+test_that("Creating cubes from BDC - LANDSAT per tile", {
+    # check "BDC_ACCESS_KEY" - mandatory one per user
+    bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
+    testthat::skip_if(nchar(bdc_access_key) == 0,
+                      message = "No BDC_ACCESS_KEY defined in environment."
+    )
+    tile <- "038046"
+    start_date <- "2021-05-01"
+    end_date <- "2021-09-30"
+    bands <- c("NDVI", "EVI")
+    # create a raster cube file based on the information about the files
+    bdc_l8_cube <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "LANDSAT-OLI-16D",
+            bands = bands,
+            tiles = tile,
+            start_date = start_date,
+            end_date = end_date,
+            progress = FALSE
+        )
+    },
+    .default = NULL)
 
+    testthat::skip_if(purrr::is_null(bdc_l8_cube),
+                      message = "BDC cube LANDSAT-OLI-16D is not accessible"
+    )
+    expect_equal(bdc_l8_cube$tile, tile)
+    expect_true(all(sits_bands(bdc_l8_cube) %in% bands))
+    # test timeline
+    timeline <- sits_timeline(bdc_l8_cube)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+    # test raster obj
+    r_obj <- .raster_open_rast(bdc_l8_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(bdc_l8_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating cubes from BDC - LANDSAT per roi", {
+    # check "BDC_ACCESS_KEY" - mandatory one per user
+    bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
+    testthat::skip_if(nchar(bdc_access_key) == 0,
+                      message = "No BDC_ACCESS_KEY defined in environment."
+    )
+    roi <- c(lon_min = -53.9311, lat_min = -13.2697,
+             lon_max = -53.0595, lat_max = -12.6704)
+    start_date <- "2021-05-01"
+    end_date <- "2021-09-30"
+    bands <- c("NDVI", "EVI")
+    # create a raster cube file based on the information about the files
+    bdc_l8_cube <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "LANDSAT-OLI-16D",
+            bands = bands,
+            roi = roi,
+            start_date = start_date,
+            end_date = end_date,
+            progress = FALSE
+        )
+    },
+    .default = NULL)
+
+    testthat::skip_if(purrr::is_null(bdc_l8_cube),
+                      message = "BDC cube LANDSAT-OLI-16D is not accessible"
+    )
+    expect_true(all(sits_bands(bdc_l8_cube) %in% bands))
+    bbox_cube <- sits_bbox(bdc_l8_cube, as_crs = "EPSG:4326")
+    intersects <- .cube_intersects(bdc_l8_cube, roi)
+    expect_true(all(intersects))
+    # test timeline
+    timeline <- sits_timeline(bdc_l8_cube)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+    # test raster obj
+    r_obj <- .raster_open_rast(bdc_l8_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(bdc_l8_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating cubes from BDC - SENTINEL-2 - roi", {
+    # check "BDC_ACCESS_KEY" - mandatory one per user
+    bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
+    testthat::skip_if(nchar(bdc_access_key) == 0,
+                      message = "No BDC_ACCESS_KEY defined in environment."
+    )
+    roi <- c(lon_min = -53.9311, lat_min = -13.2697,
+             lon_max = -53.0595, lat_max = -12.6704)
+    start_date <- "2021-05-01"
+    end_date <- "2021-09-30"
+    bands <- c("NDVI", "EVI")
+    # create a raster cube
+    bdc_s2_cube <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "SENTINEL-2-16D",
+            bands = bands,
+            roi = roi,
+            start_date = start_date,
+            end_date = end_date,
+            progress = FALSE
+        )
+    },
+    .default = NULL)
+    testthat::skip_if(purrr::is_null(bdc_s2_cube),
+                      message = "BDC cube SENTINEL-2-16D is not accessible"
+    )
+    expect_true(all(sits_bands(bdc_s2_cube) %in% c("NDVI", "EVI")))
+    bbox_cube <- sits_bbox(bdc_s2_cube, as_crs = "EPSG:4326")
+    intersects <- .cube_intersects(bdc_s2_cube, roi)
+    expect_true(all(intersects))
+    # test timeline
+    timeline <- sits_timeline(bdc_s2_cube)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+    # test raster obj
+    r_obj <- .raster_open_rast(bdc_s2_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(bdc_s2_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
+test_that("Creating cubes from BDC - SENTINEL-2 - tile", {
+    # check "BDC_ACCESS_KEY" - mandatory one per user
+    bdc_access_key <- Sys.getenv("BDC_ACCESS_KEY")
+    testthat::skip_if(nchar(bdc_access_key) == 0,
+                      message = "No BDC_ACCESS_KEY defined in environment."
+    )
+    start_date <- "2021-05-01"
+    end_date <- "2021-09-30"
+    bands <- c("NDVI", "EVI")
+    tiles <- "021019"
+    # create a raster cube file
+    bdc_s2_cube_t <- .try({
+        sits_cube(
+            source = "BDC",
+            collection = "SENTINEL-2-16D",
+            bands = bands,
+            tiles = tiles,
+            start_date = start_date,
+            end_date =  end_date,
+            progress = FALSE
+        )
+    },
+    .default = NULL)
+
+    testthat::skip_if(purrr::is_null(bdc_s2_cube_t),
+                      message = "BDC cube SENTINEL-2-16D is not accessible"
+    )
+    expect_true(all(sits_bands(bdc_s2_cube_t) %in% c("NDVI", "EVI")))
+    # test timeline
+    timeline <- sits_timeline(bdc_s2_cube_t)
+    expect_true(timeline[1] <= as.Date(start_date))
+    expect_true(timeline[length(timeline)] <= as.Date(end_date))
+    # test raster obj
+    r_obj <- .raster_open_rast(bdc_s2_cube_t$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(bdc_s2_cube_t)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
+})
 test_that("Creating cubes from DEA", {
-
+    # try to create a DEA cube
     dea_cube <- .try({
         sits_cube(
             source = "DEAFRICA",
@@ -246,22 +407,15 @@ test_that("Creating cubes from DEA", {
     },
     .default = NULL
     )
-
     testthat::skip_if(purrr::is_null(dea_cube),
         message = "DEAFRICA is not accessible"
     )
-
     expect_true(all(sits_bands(dea_cube) %in% c("B01", "B04", "B05")))
-
     r <- .raster_open_rast(.tile_path(dea_cube))
-
     expect_equal(dea_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
     expect_equal(dea_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
 })
-
 test_that("Creating cubes from DEA - error using tiles", {
-
-
     expect_error(
         object = {
             dea_cube <-
@@ -278,7 +432,6 @@ test_that("Creating cubes from DEA - error using tiles", {
         regexp = "DEAFRICA cubes do not support searching for tiles"
     )
 })
-
 test_that("Creating Sentinel cubes from MPC", {
     s2_cube <- .try({
         sits_cube(
@@ -293,26 +446,21 @@ test_that("Creating Sentinel cubes from MPC", {
     },
     .default = NULL
     )
-
-    testthat::skip_if(
-        purrr::is_null(s2_cube),
+    testthat::skip_if(purrr::is_null(s2_cube),
         "MPC is not accessible"
     )
-
     expect_true(all(sits_bands(s2_cube) %in% c("B05", "CLOUD")))
-
     r <- .raster_open_rast(.tile_path(s2_cube))
-
     expect_equal(s2_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
     expect_equal(s2_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
+    r_obj <- .raster_open_rast(s2_cube$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(s2_cube)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
 })
-
 test_that("Creating Sentinel cubes from MPC with ROI", {
-
     roi <- c(lon_min = -48.28579, lat_min = -16.05026,
              lon_max = -47.30839, lat_max = -15.50026)
-
-    s2_cube <- .try({
+    s2_cube_mpc <- .try({
         sits_cube(
             source = "MPC",
             collection = "SENTINEL-2-L2A",
@@ -325,22 +473,49 @@ test_that("Creating Sentinel cubes from MPC with ROI", {
     },
     .default = NULL
     )
-
-    testthat::skip_if(purrr::is_null(s2_cube), "MPC is not accessible")
-
-    expect_true(all(sits_bands(s2_cube) %in% c("B05", "CLOUD")))
-    expect_equal(nrow(s2_cube), 3)
-
-    bbox_cube <- sits_bbox(s2_cube, as_crs = "EPSG:4326")
-    bbox_cube_1 <- sits_bbox(.tile(s2_cube), as_crs = "EPSG:4326")
+    testthat::skip_if(purrr::is_null(s2_cube_mpc), "MPC is not accessible")
+    expect_true(all(sits_bands(s2_cube_mpc) %in% c("B05", "CLOUD")))
+    expect_equal(nrow(s2_cube_mpc), 3)
+    bbox_cube <- sits_bbox(s2_cube_mpc, as_crs = "EPSG:4326")
+    bbox_cube_1 <- sits_bbox(.tile(s2_cube_mpc), as_crs = "EPSG:4326")
     expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
     expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
-})
+    r_obj <- .raster_open_rast(s2_cube_mpc$file_info[[1]]$path[1])
+    cube_nrows <- .tile_nrows(s2_cube_mpc)
+    expect_true(.raster_nrows(r_obj) == cube_nrows)
 
-test_that("Creating Harmonized Landsat Sentinel cubes from HLS", {
+})
+test_that("Creating LANDSAT cubes from MPC with ROI", {
     roi <- c(lon_min = -48.28579, lat_min = -16.05026,
              lon_max = -47.30839, lat_max = -15.50026)
+    l8_cube_mpc <- .try({
+        sits_cube(
+            source = "MPC",
+            collection = "LANDSAT-C2-L2",
+            roi = roi,
+            bands = c("NIR08", "CLOUD"),
+            start_date = as.Date("2018-07-18"),
+            end_date = as.Date("2018-08-23"),
+            progress = FALSE
+        )
+    },
+    .default = NULL
+    )
+    testthat::skip_if(purrr::is_null(l8_cube_mpc), "MPC is not accessible")
+    expect_true(all(sits_bands(l8_cube_mpc) %in% c("NIR08", "CLOUD")))
+    expect_equal(nrow(l8_cube_mpc), 2)
+    bbox_cube <- sits_bbox(l8_cube_mpc, as_crs = "EPSG:4326")
+    bbox_cube_1 <- sits_bbox(.tile(l8_cube_mpc), as_crs = "EPSG:4326")
+    expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
+    expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
+    r_obj <- .raster_open_rast(l8_cube_mpc$file_info[[1]]$path[1])
+    tile_nrows <- .tile_nrows(l8_cube_mpc)[[1]]
+    expect_true(.raster_nrows(r_obj) == tile_nrows)
 
+})
+test_that("Creating Harmonized Landsat Sentinel HLSS30 cubes", {
+    roi <- c(lon_min = -48.28579, lat_min = -16.05026,
+             lon_max = -47.30839, lat_max = -15.50026)
     hls_cube_s2 <- .try({
         sits_cube(
             source = "HLS",
@@ -354,18 +529,19 @@ test_that("Creating Harmonized Landsat Sentinel cubes from HLS", {
     },
     .default = NULL
     )
-
     testthat::skip_if(
         purrr::is_null(hls_cube_s2),
         "HLSS30 collection is not accessible"
     )
-
     expect_true(all(sits_bands(hls_cube_s2) %in%
                         c("GREEN", "NIR-NARROW", "SWIR-1", "CLOUD")))
     expect_true(all(hls_cube_s2$satellite == "SENTINEL-2"))
     expect_true(all(hls_cube_s2$tile %in% c("23LKC", "22LHH", "22LGH")))
     expect_true(all(.fi(hls_cube_s2)$xres == 30))
     expect_true(all(.fi(hls_cube_s2)$yres == 30))
+    r_obj <- .raster_open_rast(hls_cube_s2$file_info[[1]]$path[1])
+    tile_nrows <- .tile_nrows(hls_cube_s2)[[1]]
+    expect_true(.raster_nrows(r_obj) == tile_nrows)
 
     hls_cube_l8 <- .try({
         sits_cube(
@@ -380,12 +556,10 @@ test_that("Creating Harmonized Landsat Sentinel cubes from HLS", {
     },
     .default = NULL
     )
-
     testthat::skip_if(
         purrr::is_null(hls_cube_l8),
         "HLSL30 collection is not accessible"
     )
-
     expect_true(all(sits_bands(hls_cube_l8) %in%
                         c("GREEN", "NIR-NARROW", "SWIR-1", "CLOUD")))
     expect_true(all(hls_cube_l8$satellite == "LANDSAT-8"))
@@ -394,7 +568,6 @@ test_that("Creating Harmonized Landsat Sentinel cubes from HLS", {
     expect_true(all(.fi(hls_cube_l8)$yres == 30))
 
     hls_cube_merge <- sits_merge(hls_cube_s2, hls_cube_l8)
-
     merge_23LKC <- dplyr::filter(hls_cube_merge, tile == "23LKC")
     s2_23LKC <- dplyr::filter(hls_cube_s2, tile == "23LKC")
     l8_23LKC <- dplyr::filter(hls_cube_l8, tile == "23LKC")
@@ -416,43 +589,39 @@ test_that("Creating Sentinel cubes from AWS", {
     },
     .default = NULL
     )
-
     testthat::skip_if(
         purrr::is_null(s2_cube),
         "AWS is not accessible"
     )
-
     expect_true(all(sits_bands(s2_cube) %in% c("B05", "CLOUD")))
-
     r <- .raster_open_rast(.tile_path(s2_cube))
-
     expect_equal(s2_cube$xmax[[1]], .raster_xmax(r), tolerance = 1)
     expect_equal(s2_cube$xmin[[1]], .raster_xmin(r), tolerance = 1)
-
-    s2_cube_aws1 <- .try({
+})
+test_that("Creating LANDSAT cubes from AWS with ROI", {
+    roi <- c(lon_min = -48.28579, lat_min = -16.05026,
+             lon_max = -47.30839, lat_max = -15.50026)
+    l8_cube_aws <- .try({
         sits_cube(
             source = "AWS",
-            collection = "SENTINEL-S2-L2A-COGS",
-            tiles = "20LKP",
-            bands = c("B05"),
-            start_date = "2023-04-01",
-            end_date = "2023-04-20",
+            collection = "LANDSAT-C2-L2",
+            roi = roi,
+            bands = c("NIR08", "CLOUD"),
+            start_date = as.Date("2018-07-18"),
+            end_date = as.Date("2018-08-23"),
             progress = FALSE
         )
     },
     .default = NULL
     )
-    testthat::skip_if(
-        purrr::is_null(s2_cube_aws1),
-        "AWS is not accessible"
-    )
-
-    expect_true(all(sits_bands(s2_cube_aws1) %in% c("B05", "CLOUD")))
-
-    r <- .raster_open_rast(.tile_path(s2_cube_aws1))
-
-    expect_equal(s2_cube_aws1$xmax[[1]], .raster_xmax(r), tolerance = 1)
-    expect_equal(s2_cube_aws1$xmin[[1]], .raster_xmin(r), tolerance = 1)
-
+    testthat::skip_if(purrr::is_null(l8_cube_aws), "AWS is not accessible")
+    expect_true(all(sits_bands(l8_cube_aws) %in% c("NIR08", "CLOUD")))
+    expect_equal(nrow(l8_cube_aws ), 2)
+    bbox_cube <- sits_bbox(l8_cube_aws , as_crs = "EPSG:4326")
+    bbox_cube_1 <- sits_bbox(.tile(l8_cube_aws), as_crs = "EPSG:4326")
+    expect_true(bbox_cube["xmax"] >= bbox_cube_1["xmax"])
+    expect_true(bbox_cube["ymax"] >= bbox_cube_1["ymax"])
+    r_obj <- .raster_open_rast(l8_cube_aws$file_info[[1]]$path[1])
+    tile_nrows <- .tile_nrows(l8_cube_aws)[[1]]
+    expect_true(.raster_nrows(r_obj) == tile_nrows)
 })
-