+++ tests/testthat/test-cube_copy.R
@@ -3,7 +3,7 @@ test_that("Downloading and cropping cubes from BDC", {
         {
             sits_cube(
                 source = "BDC",
-                collection = "CB4-16D-2",
+                collection = "CBERS-WFI-16D",
                 tiles = c("007004", "007005"),
                 bands = c("B15", "CLOUD"),
                 start_date = "2018-01-01",