+++ tests/testthat/test-mixture_model.R
@@ -3,7 +3,7 @@ test_that("Mixture model tests", {
     # Create a sentinel-2 cube
     s2_cube <- sits_cube(
         source = "AWS",
-        collection = "SENTINEL-S2-L2A-COGS",
+        collection = "SENTINEL-2-L2A",
         tiles = "20LKP",
         bands = c("B02", "B03", "B04", "B8A", "B11", "B12", "CLOUD"),
         start_date = "2019-07-01",