+++ tests/testthat/test-regularize.R
@@ -2,9 +2,9 @@ test_that("Regularizing cubes from AWS, and extracting samples from them", {
     s2_cube_open <- .try({
         sits_cube(
             source = "AWS",
-            collection = "SENTINEL-S2-L2A-COGS",
+            collection = "SENTINEL-2-L2A",
             tiles = c("20LKP", "20LLP"),
-            bands = c("B8A", "SCL"),
+            bands = c("B8A", "CLOUD"),
             start_date = "2018-10-01",
             end_date = "2018-11-01",
             multicores = 1,