+++ NEWS.md
@@ -1,8 +1,14 @@
 # SITS Release History
 
-# What's new in SITS version 1.4.0
+# What's new in SITS version 1.4
 
-### New features in SITS version 1.3.0
+### New features in SITS version 1.4.1
+* Updated access to collections in Brazil Data Cube, HLS, and AWS
+* Corrected errors in labelling of classified cubes
+* Created a factory of functions for segmentation
+
+
+### New features in SITS version 1.4.0
 * New function for image segementation based on `supercells` package
 * New version of `sits_get_data()` to extract average values of time series based on segments
 * Support for Harmonized Landsat Sentinel (HLS) collections from NASA
@@ -11,7 +17,7 @@
 * General big fixes
 
 
-# What's new in SITS version 1.3.0
+# What's new in SITS version 1.3
 
 ### New features in SITS version 1.3.0
 * Remove NOTES and WARNINGS pointed out by CRAN